Made GLX 1.3 a requirement for the xlib OpenGL port
Modern Unix like systems (GNU/Linux, etc...) just have to be kept
update. After all it is Free Software (aka. "Open Source") and not
some proprietary OS that forces us to drag along old workarounds for
backwards compatibillity because there is no way of changing it and
the developers wont.