Fixed WinDOS redrawing bug
When the window is made visible, the widgets weren't redrawn until the
window received an event that was not WM_PAINT, as the WM_PAINT event
simply displayed the canvas pixmap and returned without checking for
dirty rects.

For some reason, this did not occour before the build system switch.