Implemented double click detection in icon view widget
Simply stores the last clicked icon and a timestamp. If any user input
happens, other than left mouse button click/release, the icon pointer
is cleared. If the timestamp is further in the past than a certain
threshold, the icon pointer is cleared.
-> If a an icon gets clicked and the icon pointer points to the same
   icon, a select event is generated

Appears a little flaky on my system with spurious double clicks, but
that might be my broken mouse (other programs on the same system show
the same behaviour)

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>